Spiking out some changes
