+++ exercises/isogram/partial.tt
@@ -1,0 +1,1 @@
+<b>Hello I am a partial</b>
