+++ lib/isogram_cases.rb
@@ -1,16 +1,18 @@
-require_relative 'exercise_testcases'
-require_relative 'exercise_testcase'
+require_relative 'exercise_test_cases'
+require_relative 'exercise_test_case'
 
-class IsogramCases < ExerciseTestCases
-end
+class IsogramCases < ExerciseTestCases; end
 
 class IsogramCase < ExerciseTestCase
+
   def workload
     [
       "string = '#{canonical_data.input}'",
       "#{assertion} Isogram.isogram?(string), '#{failure_message}'"
     ]
   end
+
+  private
 
   def failure_message
     "#{canonical_data.input.inspect} #{is_or_isnt} an isogram"
